+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) 
 - Added validation checks for configuration
 - Added checks for argument counts for each command
 - Added limit for asset size and number of concurrent downloads
+- Added kache to the list of known tools
 
 ### Changed